Changed board element type to uint8

+++ node.py
@@ -3,7 +3,7 @@ import numpy as np
 
 class Node:
     def __init__(self, n, initial):
-        board = np.zeros((n, n), dtype=np.int32)
+        board = np.zeros((n, n), dtype=np.uint8)
         index = 0
         for i in range(n):
             for j in range(n):
@@ -17,21 +17,15 @@ class Node:
             for j in range(n):
                 child = np.array(self.v)
                 # change the touched token
-                if child[i,j]:
-                    child[i,j] = 0
+                if child[i, j]:
+                    child[i, j] = 0
                 else:
-                    child[i,j] = 1
+                    child[i, j] = 1
                 # left
-                if j-1 > 0:
+                if j - 1 > 0:
                     pass
                 # right
 
                 # top
 
                 # bottom
-
-
-
-
-
-
